+++ recipes/user.rb
@@ -2,7 +2,7 @@
 # Cookbook:: chef-splunk
 # Recipe:: user
 #
-# Copyright:: 2014-2016, Chef Software, Inc.
+# Copyright:: 2014-2020, Chef Software, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,6 +22,8 @@ group node['splunk']['user']['username'] do
   system true if %w(linux).include?(node['os'])
 end
 
+node.default['splunk']['user']['home'] = splunk_dir
+
 user node['splunk']['user']['username'] do
   comment node['splunk']['user']['comment']
   shell node['splunk']['user']['shell']
@@ -29,4 +31,5 @@ user node['splunk']['user']['username'] do
   uid node['splunk']['user']['uid']
   home node['splunk']['user']['home']
   system true if %w(linux).include?(node['os'])
+  notifies :stop, 'service[splunk]', :before if splunk_installed?
 end